Add introduction
Prompts:

Looks great, thanks! Can you add a short bit of text to the top of the application explaining to users what the tool is, why it exists and and how to use it, along with a link to the source code on GitHiub? Make it able to be collapsed/hidden and store that state in the share link also

Goo start. Please make the text default to being visible and make the UI for showing/hiding it more visual (move "Show/Hide Introduction" to accessible text and/or a tooltip and use a visual element for toggling the visibility)