Add shareable links (#4)
Squash merge of feature/share

Adds the ability to share the state of the app via a URL (by serializing the
table state and contents to the URL hash) as well as a CSV export function.

This is a squash merge of multiple AI generated, human modified commits.
Prompts are in the commit messages from the  original branch.

CSV import was attempted, but did not work and was removed.

Added an intro paragraph describing the application and linking to the source.

Small look and feel improvements.