Add a dns domain column (#7)
* Add a DNS Domain column

Prompt:

Modify the data model to add a new column for "dns domain".
This means the table will need to be adjusted to have a new column,
and the JSON import/export will need to be updated, and the CSV
export will need the new column. Place the new column after
"status" and before "description"

* Cleanup

* Don't map missing dns_domain, that places it as the first field