Add Markdown export support (#8)
Initial Markdown export support

Prompt:

It would be nice to have a way to export Markdown from the CIDR sheet.
The Markdown should list the entire contents of the table as a Markdown table.
Bonus points if this is in a dedicated function which could be run outside
of the browser, I.E. in a CI/CD pipeline to produce a Markdown version of any
JSON files, and we can call that function from the "Export/Share" button but
also it could be pasted in to a separate javascript file and run separately
via node.js (loading in a JSON file).